--- Assets/Prefabs/GameManager.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7dfddf2a70036c64293dd9ab5538b0ef
-timeCreated: 1485006118
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
