+++ Assets/Scripts/LetterCollider.cs
@@ -4,13 +4,21 @@ using UnityEngine;
 using UnityEngine.UI;
 
 public class LetterCollider : MonoBehaviour {
-	private GameObject boundLetter;
-	public GameObject BoundLetter {
-		set { boundLetter = value; }
+	public GameObject letter;
+	private GameObject instanciedLetter;
+	public string value;
+	private GameObject parent;
+
+	void Start () {
+		parent = GameObject.FindGameObjectWithTag ("Canvas");
+		// Instanciation d'une lettre
+		Vector3 pos = parent.transform.position;
+		instanciedLetter = Instantiate (letter, transform.position, new Quaternion (), parent.transform);
+		instanciedLetter.GetComponent<Text>().text = value;
 	}
 
 	void OnTriggerEnter2D(Collider2D other) {
-		Destroy (boundLetter);
+		Destroy (instanciedLetter);
 		Destroy (gameObject);
 	}
 }
