--- Assets/Scripts/LetterManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f265e1e65b8428d49ab6c21991d72093
-timeCreated: 1484988968
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
